# openapi.yaml
openapi: 3.0.3
info:
  title: Guardian Mobile API
  description: The remote anti-theft system API for Android phones. Endpoints marked as "(planned)" are not yet implemented.
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CommandDto:
      type: object
      required:
        - commandType
        - targetDeviceId
      properties:
        commandType:
          type: string
          enum: [RING, LOCATE, LOCK, SHOW_MESSAGE, VIBRATE, THEFT_MODE, WIPE]
        targetDeviceId:
          type: string
          format: uuid
        payload:
          type: object
          description: "Optional data like a custom message to show, or coordinates."
...
security:
  - bearerAuth: []
paths:
...
  /devices/{deviceId}/commands:
    post:
      summary: (planned) Send a remote command to a protected device
      tags:
        - Commands
      parameters:
        - name: deviceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommandDto'
      responses:
        '202':
          description: Command accepted and queued/sent
        '403':
          description: Forbidden (Command ownership check failed)
...